Fix issue causing multiple instances of link tags to be inserted.
If you have a large css code base which import from eachother, webpack will
send a single HMR update with an entry for each of the underlying source files.

Without this patch, if you are compiling all the css into a single bundle,
css-hot-reloader will insert a new copy of the bundle for each underlying module
in the change tree that maps to said bundle. Result: memory usage goes through the
roof and possible browser crash.